% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_authentication.R
\name{set_api_token}
\alias{set_api_token}
\title{Set a Laws.Africa API token}
\usage{
set_api_token(api_token = NULL)
}
\arguments{
\item{api_token}{Laws.Africa user account API token.}
}
\value{
User's API token recorded into system environment as
\emph{LAWS_AFRICA_TOKEN}.
}
\description{
Set a Laws.Africa API token
}
\examples{
if (FALSE) {
  set_api_token(api_token = "abcd")
}

}
