% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africalaws.R
\docType{package}
\name{africalaws}
\alias{africalaws-package}
\alias{africalaws}
\title{Interface to the Laws.Africa API}
\description{
Laws.Africa \url{https://laws.africa} endeavours to facilitate African
governments in offering sustainable, free access to dependable digital laws.
It aims to achieve this by ensuring that the laws are easily accessible,
user-friendly, educational, and reusable. This initiative seeks to empower
citizens with knowledge of their legal rights and obligations while
promoting transparency and accountability within the legal system.
Laws.Africa offers a content application programming interface (API) to fetch
legislative content and metadata. The API is a read-only API for listing and
fetching published versions of legislative works. This package interfaces
with this API to allow access using R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OxfordIHTM/africalaws}
  \item \url{http://oxford-ihtm.io/africalaws/}
  \item Report bugs at \url{https://github.com/OxfordIHTM/africalaws/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jillian Francise Lee [copyright holder]
  \item Sanjeev Pugazhendhi [copyright holder]
  \item Proochista Ariana [copyright holder]
}

}
\keyword{internal}
