% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_countries.R
\name{al_get_countries}
\alias{al_get_countries}
\title{Retrieve countries}
\usage{
al_get_countries(base_url = "https://api.laws.africa/v3/", flatten = TRUE)
}
\arguments{
\item{base_url}{Base URL for the Laws.Africa API. Current default set to
https://api.laws.africa/v3/.}

\item{flatten}{Logical. Should the output JSON be flattened into a tibble.
Defeault is TRUE.}
}
\value{
If \code{flatten = TRUE}, a named list of two tibbles the first of which
is for \emph{countries} information and the second for \emph{localities} information.
Otherwise, a list output created from the JSON output of the request.
}
\description{
Retrieve countries
}
\examples{
al_get_countries()

}
