% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_countries.R, R/al_works.R
\name{al_get_countries}
\alias{al_get_countries}
\alias{al_get_works}
\title{Retrieve countries}
\usage{
al_get_countries(base_url = "https://api.laws.africa/v3/", flatten = TRUE)

al_get_works(country, base_url = "https://api.laws.africa/v3/", flatten = TRUE)
}
\arguments{
\item{base_url}{Base URL for the Laws.Africa API. Current default set to
https://api.laws.africa/v3/.}

\item{flatten}{Logical. Should the output JSON be flattened into a tibble.
Defeault is TRUE.}

\item{country}{Two-letter ISO 3166-1 alpha-2 country code.}
}
\value{
If \code{flatten = TRUE}, a named list of two tibbles the first of which
is for \emph{countries} information and the second for \emph{localities} information.
Otherwise, a list output created from the JSON output of the request.

If \code{flatten = TRUE}, a tibble of works from specified country.
Otherwise, a list output created from the JSON output of the request.
}
\description{
A \emph{Work} is a piece of legislation, such as an act, regulation or by-law. A
work may be amended over time and may even have its title changed. A work is
uniquely identified by a work FRBR URI which never changes.
}
\examples{
if (Sys.getenv("LAWS_AFRICA_TOKEN") != "") al_get_countries()

if (Sys.getenv("LAWS_AFRICA_TOKEN") != "") al_get_countries()
  al_get_works(country = "za-cpt")

}
